Feature/math and token tracking (#21)
* Adding and removing token logic added / fixed
* Improved math, precision, etc.

+++ programs/dtfs/src/instructions/trade/approve_trade.rs
@@ -79,8 +79,8 @@ pub fn handler(
     trade_id: u64,
     sell_limit: Range,
     buy_limit: Range,
-    start_price: u64,
-    end_price: u64,
+    start_price: u128,
+    end_price: u128,
     ttl: u64,
 ) -> Result<()> {
     ctx.accounts.validate()?;
